[Silver IV] Title: 점화식, Time: 0 ms, Memory: 13204 KB -BaekjoonHub

+++ 백준/Silver/13699. 점화식/점화식.rs
@@ -2,15 +2,15 @@ use std::io::{self, stdin};
 
 fn main() {
     let buf = io::read_to_string(stdin()).unwrap();
-    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<u128>);
+    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<usize>);
     let n = input.next().unwrap();
-    let mut dp = [0u128; 36];
+    let mut dp = [0usize; 36];
     dp[0] = 1;
     dp[1] = 1;
     for i in 2..=n {
         for j in 0..i {
-            dp[i as usize] += dp[j as usize] * dp[(i - 1 - j) as usize];
+            dp[i] += dp[j] * dp[i - 1 - j];
         }
     }
-    print!("{}", dp[n as usize]);
+    print!("{}", dp[n]);
 }
